public/Backend/requirements.txt
fastapi
uvicorn
sqlalchemy
asyncpg
psycopg2-binary
passlib[bcrypt]
python-jose
pydantic